move chain + config + db to /chainlink
