+++ relayer/pkg/chainlink/chain/chain.go
@@ -4,9 +4,10 @@ import (
 	"context"
 
 	"github.com/smartcontractkit/chainlink-relay/pkg/types"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/config"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/db"
 	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/txm"
 	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/starknet"
-	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/starknet/db"
 
 	// unused module to keep in go.mod and prevent ambiguous import
 	_ "github.com/btcsuite/btcd/chaincfg/chainhash"
@@ -21,7 +22,7 @@ type ChainSet interface {
 type Chain interface {
 	types.Service
 
-	Config() starknet.Config
+	Config() config.Config
 	UpdateConfig(*db.ChainCfg)
 
 	TxManager() txm.TxManager
