+++ relayer/pkg/chainlink/ocr2/contract_cache.go
@@ -8,10 +8,9 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/starknet"
-
 	"github.com/smartcontractkit/chainlink-relay/pkg/logger"
 	"github.com/smartcontractkit/chainlink-relay/pkg/utils"
+	"github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/config"
 	"github.com/smartcontractkit/libocr/offchainreporting2/types"
 )
 
@@ -32,11 +31,11 @@ type contractCache struct {
 	stop, done chan struct{}
 
 	reader *contractReader
-	cfg    starknet.Config
+	cfg    config.Config
 	lggr   logger.Logger
 }
 
-func NewContractCache(cfg starknet.Config, reader *contractReader, lggr logger.Logger) *contractCache {
+func NewContractCache(cfg config.Config, reader *contractReader, lggr logger.Logger) *contractCache {
 	return &contractCache{
 		cfg:    cfg,
 		reader: reader,
